add some annotations and move math function from util to mathExtensions
